requirements.txt
# requirements.txt — versões atualizadas e compatíveis com Python 3.14
fastapi>=0.115.6
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
openai>=1.54.0
mcp>=1.2.0
pydantic-settings>=2.6.1
httpx>=0.28.0
starlette>=0.41.0